added /f map|rename

+++ CodexFactions/src/org/codex/factions/executors/Disbander.java
@@ -44,7 +44,7 @@ public class Disbander implements Execute {
 			fac.removeClaim(c);
 		}
 
-		for(int i: fac.getRelations().keySet()) {
+		for(byte i: fac.getRelations().keySet()) {
 			for(String facName: fac.getRelations().get(i)) {
 				FactionObject fac2 = FactionsMain.getFactionFromName(facName);
 				fac2.addRelationship(fac, Relationship.NEUTRAL);
